requirements.txt
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
py4j==0.10.9.7
pyspark==3.5.5
pytest==8.3.5
PyYAML==6.0.2
